Add a wordLang-level static data mechanism
wordLang is extended with a StaticRead instruction, and the semantics
state has a single (currently read only) "rodata" array.  This is
handled in word_to_stack by appending it to the stack-level bitmaps, and
recording the offset of the rodata in a new StaticOffset store field,
which is initialized by stack_remove.  The actual bitmaps stay at offset
0 in order to allow more efficient instructions to be used for setting
up stack frames.

This is somewhat of an interim solution; I beleive we should eventually
support multiple read-only and read-write static data objects,
accessible via pointers using the Loc mechanism or something similar.