dataLang support for word-at-a-time strings
This adds a dataLang primitive for initializing strings and lowers it to
a sequence of _word_ writes with a precomputed pointer, resulting in
substantially fewer instructions than the previous lowering using byte
writes (on x64, 2 per 8 bytes instead of 56 per 8 bytes).  Modifies the
memory proofs to allow for writing whole strings at once, replaces
write_bytes with a translatable version, and adds plumbing to support
our first endian-specific transformation.