Compile strings to read-only data
This adds the actual extraction of long strings from programs in
data_to_word, adds a StringLitLoop helper which copies rodata to byte
objects, and changes data_to_word to use the helper instead of
populating long strings with unrolled code.

Also changes the String op to have nonzero stack usage, passes the
endian flag to strings_to_rodata, and fixes a bug in compilationLib with
bitmap entries not representable as a uint.

yesProofScript is broken by this version of the commit.

Fixes #773.